requirements.txt
python-frontmatter
markata
markupsafe==2.0.1
rich
textual
